+++ src/modules/ChannelSettings/components/ModerationPanel/index.tsx
@@ -24,13 +24,13 @@ import MutedMemberList from './MutedMemberList';
 
 import { useChannelSettingsContext } from '../../context/ChannelSettingsProvider';
 
-const kFormatter = (num: number): string|number => {
+const kFormatter = (num: number): string | number => {
   return Math.abs(num) > 999
     ? `${(Math.abs(num) / 1000).toFixed(1)}K`
     : num;
 };
 
-export default function AdminPannel(): ReactElement {
+export default function ModerationPanel(): ReactElement {
   const [frozen, setFrozen] = useState(false);
 
   const { stringSet } = useContext(LocalizationContext);