+++ src/modules/GroupChannel/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { GroupChannelProvider, GroupChannelProviderProps } from './context/GroupChannelProvider';
 import GroupChannelUI, { GroupChannelUIProps } from './components/GroupChannelUI';
 
-export interface GroupChannelProps extends GroupChannelProviderProps, GroupChannelUIProps {}
+export interface GroupChannelProps extends GroupChannelProviderProps, GroupChannelUIProps { }
 export const GroupChannel = (props: GroupChannelProps) => {
   return (
     <GroupChannelProvider {...props}>