+++ src/ui/MessageContent/MessageBody/index.tsx
@@ -21,6 +21,7 @@ import { Nullable, SendbirdTheme } from '../../../types';
 import { GroupChannel } from '@sendbird/chat/groupChannel';
 import { match } from 'ts-pattern';
 import TemplateMessageItemBody from '../../TemplateMessageItemBody';
+import type { OnBeforeDownloadFileMessageType } from '../../../modules/GroupChannel/context/GroupChannelProvider';
 
 const MESSAGE_ITEM_BODY_CLASSNAME = 'sendbird-message-content__middle__message-item-body';
 export type RenderedTemplateBodyType = 'failed' | 'composite' | 'simple';
@@ -29,8 +30,9 @@ export interface MessageBodyProps {
   channel: Nullable<GroupChannel>;
   message: CoreMessageType;
   showFileViewer?: (bool: boolean) => void;
-  onMessageHeightChange?: (isBottomMessageAffected?: boolean) => void;
   onTemplateMessageRenderedCallback?: (renderedTemplateBodyType: RenderedTemplateBodyType) => void;
+  onMessageHeightChange?: () => void;
+  onBeforeDownloadFileMessage?: OnBeforeDownloadFileMessageType;
 
   mouseHover: boolean;
   isMobile: boolean;
@@ -46,6 +48,7 @@ export default function MessageBody(props: MessageBodyProps): ReactElement {
     showFileViewer,
     onMessageHeightChange,
     onTemplateMessageRenderedCallback,
+    onBeforeDownloadFileMessage,
 
     mouseHover,
     isMobile,
@@ -102,6 +105,7 @@ export default function MessageBody(props: MessageBodyProps): ReactElement {
         isByMe={isByMe}
         mouseHover={mouseHover}
         isReactionEnabled={isReactionEnabledInChannel}
+        onBeforeDownloadFileMessage={onBeforeDownloadFileMessage}
       />
     ))
     .when(isMultipleFilesMessage, () => (
@@ -113,6 +117,7 @@ export default function MessageBody(props: MessageBodyProps): ReactElement {
         isReactionEnabled={isReactionEnabledInChannel}
         threadMessageKindKey={threadMessageKindKey}
         statefulFileInfoList={statefulFileInfoList}
+        onBeforeDownloadFileMessage={onBeforeDownloadFileMessage}
       />
     ))
     .when(isVoiceMessage, () => (