+++ src/ui/UserListItem/index.tsx
@@ -13,7 +13,7 @@ import ContextMenu, { MenuItems } from '../ContextMenu';
 import Label, { LabelTypography, LabelColors } from '../Label';
 
 export interface UserListItemProps {
-  user: User;
+  user: User | Member;
   className?: string;
   checked?: boolean;
   checkBox?: boolean;
@@ -47,6 +47,7 @@ export default function UserListItem({
   onClick,
   onUserAvatarClick,
 }: UserListItemProps): ReactElement {
+  const operator = isOperator ?? (user as Member)?.role === 'operator';
   const uniqueKey = user.userId;
   const actionRef = React.useRef(null);
   const parentRef = React.useRef(null);
@@ -71,7 +72,7 @@ export default function UserListItem({
           <Avatar
             className="sendbird-user-list-item__avatar"
             ref={avatarRef}
-            src={user.profileUrl}
+            src={user?.profileUrl || user?.plainProfileUrl || ''}
             width={avatarSize}
             height={avatarSize}
             onClick={() => {
@@ -116,11 +117,9 @@ export default function UserListItem({
         color={LabelColors.ONBACKGROUND_1}
       >
         {user.nickname || stringSet.NO_NAME}
-        {
-          (currentUser === user.userId) && (
-            ' (You)'
-          )
-        }
+        {(currentUser === user.userId) && (
+          stringSet.CHANNEL_SETTING__MEMBERS__YOU
+        )}
       </Label>
       { // if there is now nickname, display userId
         !user.nickname && (
@@ -149,7 +148,7 @@ export default function UserListItem({
         )
       }
       {
-        isOperator && (
+        operator && (
           <Label
             className={[
               'sendbird-user-list-item__operator',